# perf(build): remove `--export-dynamic` flag in wasi compilation