@Brooooooklyn 2025-03-19
It was from https://github.com/toyobayashi/emnapi?tab=readme-ov-file#usage, maybe @toyobayashi know why there is `--export-dynamic` flag